timetable_project/requirements.txt
streamlit
pytesseract
pdf2image
pillow
opencv-python
pandas
numpy
python-dateutil
rapidfuzz
scikit-learn
ultralytics
ics
pytz
